pyproject.toml
[project]
name = "damnvulnerableadk"
version = "0.1.0"
description = "Basic ADK integration with sql injection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "httpx>=0.28.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
